move peek projection function to visualization module
